backend/requirements.txt
# Python 3.11 Compatible Packages
fastapi==0.104.1
uvicorn[standard]==0.23.2
pymongo>=4.6.0,<5.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6